Don't set mtime to 0 on tar files.
For some reason, every file added in the pkg_tar rule gets its modification time set to zero. This means that in addition to losing its modification time, every file produces a warning when it's untarred:
implausibly old time stamp 1969-12-31 19:00:00

This seems like a bug, but since it's done intentionally everywhere, perhaps there's a reason for it that I'm not aware of.